% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areacode.R
\docType{data}
\name{areacode}
\alias{areacode}
\title{Korean administative area and code}
\format{An object of class \code{data.frame} with 17 rows and 3 columns.}
\usage{
areacode
}
\description{
A dataset
}
\details{
@format A data.frame with 16 rows and 3 columns
 \describe{
    \item{code}{Korean administative area code}
    \item{name}{Korean administative area, level 1}
    \item{name2}{abbreviation of area}
}
}
\keyword{datasets}

