% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areacode.R
\name{changeCode}
\alias{changeCode}
\title{Change hangul code of colnames and columns of a given data.frame}
\usage{
changeCode(popdata, from = "UTF-8", to = "CP949", change.colnames = TRUE,
  cols = NULL)
}
\arguments{
\item{popdata}{A data.frame}

\item{from}{from A character string describing the current encoding.}

\item{to}{A character string describing the target encoding.}

\item{change.colnames}{If true, change the code of colnames of the given data.frame}

\item{cols}{A integer vector specify the columns to be converted}
}
\value{
A converted data.frame
}
\description{
Change hangul code of colnames and columns of a given data.frame
}
\examples{
changeCode(areacode,cols=2:3)

}

