% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areacode.R
\docType{data}
\name{kormap1}
\alias{kormap1}
\title{Korean administative map data(2014) level 1}
\format{An object of class \code{data.frame} with 8831 rows and 14 columns.}
\usage{
kormap1
}
\description{
A dataset
}
\details{
@format A data.frame with 8831 rows and 11 columns
 \describe{
    \item{id}{Korean administative area code}
    \item{long}{longitude}
    \item{lat}{latitude}
    \item{order}{An integer}
    \item{hole}{A logical}
    \item{piece}{Factor with 102 levels}
    \item{group}{Factor with 186 levels}
    \item{SP_ID}{shape id}
    \item{SIDO_CD}{sido code}
    \item{SIDO_NM}{sido name in euc-kr}
    \item{BASE_YEAR}{2014}
    \item{name}{name in utf-8}
    \item{name1}{name in euc-kr}
    \item{region}{Region}
    \item{code}{code}
}
}
\keyword{datasets}

