% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areacode.R
\docType{data}
\name{kormap2}
\alias{kormap2}
\title{Korean administative map data(2014) level 2}
\format{An object of class \code{data.frame} with 8831 rows and 14 columns.}
\usage{
kormap2
}
\description{
A dataset
}
\details{
@format A data.frame with 20397 rows and 11 columns
 \describe{
    \item{id}{Korean administative area code}
    \item{long}{longitude}
    \item{lat}{latitude}
    \item{order}{An integer}
    \item{hole}{A logical}
    \item{piece}{Factor with 37 levels}
    \item{group}{Factor with 401 levels}
    \item{SP_ID}{shape id}
    \item{base_year}{2014}
    \item{sigungu_cd}{sigungu code}
    \item{sigungu_name}{sigungu name in euc-kr}
    \item{FID}{An integer}
    \item{name}{name in utf-8}
    \item{name1}{name in euc-kr}
    \item{region}{Region}
    \item{code}{code}
}
}
\keyword{datasets}

