% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areacode.R
\docType{data}
\name{kormap3}
\alias{kormap3}
\title{Korean administative map data(2014) level 3}
\format{An object of class \code{data.frame} with 8831 rows and 14 columns.}
\usage{
kormap3
}
\description{
A dataset
}
\details{
@format A data.frame with 73474 rows and 11 columns
 \describe{
    \item{id}{Korean administative area code}
    \item{long}{longitude}
    \item{lat}{latitude}
    \item{order}{An integer}
    \item{hole}{A logical}
    \item{piece}{Factor with 11 levels}
    \item{group}{Factor with 3593 levels}
    \item{SP_ID}{shape id}
    \item{base_year}{2014}
    \item{adm_dr_cd}{dong name code}
    \item{adm_dr_nm}{dong name in euc-kr}
    \item{FID}{An integer}
    \item{name}{name in utf-8}
    \item{name1}{name in euc-kr}
    \item{region}{Region}
    \item{code}{code}
}
}
\keyword{datasets}

