% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areacode.R
\docType{data}
\name{korpop1}
\alias{korpop1}
\title{Korean population census data(2015) level 1}
\format{An object of class \code{data.frame} with 17 rows and 25 columns.}
\usage{
korpop1
}
\description{
A dataset
}
\details{
@format A data.frame with 17 rows and 25 columns
}
\keyword{datasets}

