% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areacode.R
\docType{data}
\name{korpop2}
\alias{korpop2}
\title{Korean population census data(2015) level 2}
\format{An object of class \code{data.frame} with 306 rows and 25 columns.}
\usage{
korpop2
}
\description{
A dataset
}
\details{
@format A data.frame with 306 rows and 25 columns
}
\keyword{datasets}

