% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areacode.R
\docType{data}
\name{korpop3}
\alias{korpop3}
\title{Korean population census data(2015) level 3}
\format{An object of class \code{data.frame} with 3492 rows and 25 columns.}
\usage{
korpop3
}
\description{
A dataset
}
\details{
@format A data.frame with 3492 rows and 25 columns
}
\keyword{datasets}

