% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areacode.R
\docType{data}
\name{tbc}
\alias{tbc}
\title{New patients with tubeculsosis in South Korea}
\format{An object of class \code{data.frame} with 255 rows and 5 columns.}
\usage{
tbc
}
\description{
A dataset
}
\details{
@format A long-form data.frame with 255 rows and 5 columns
 \describe{
    \item{name1}{abbreviation of area}
    \item{code}{Korean administative area code}
    \item{name}{Korean administative area, level 1}
    \item{year}{year}
    \item{NewPts}{Number of new patients}
}
}
\keyword{datasets}

